Changed the way external plugins are registered in javascript.
It's better to use the three parameter variant of
``CKEDITOR.plugins.addExternal`` because this will make the base URL
for the plugin resources available within the plugin code as
``this.path`` correctly.

Otherwise ``this.path`` contains the full URL to the plugin script
itself which consequently breaks loading of related resources
(images, translations, etc).

Before, this.path was something like

  http://server.com/++resource++my.package/ckeditor/plugin.js?t=xxxx

and now it gets correctly set to

  http://server.com/++resource++my.package/ckeditor/